refactor(library): Extract component to represent local elections
