+++ tests/test_sopn_publish_date.py
@@ -90,29 +90,6 @@ def test_publish_date_police_and_crime_commissioner():
     assert publish_date == date(2016, 4, 8)
 
 
-# Reference election: local.highland.wester-ross-strathpeffer-and-lochalsh.by.2018-12-06
-def test_publish_date_scottish_local():
-    publish_date = sopn_publish_date.local(date(2018, 12, 6), country=Country.SCOTLAND)
-
-    assert publish_date == date(2018, 11, 2)
-
-
-# Reference election: local.belfast.balmoral.2019-05-02
-def test_publish_date_northern_ireland_local():
-    publish_date = sopn_publish_date.local(
-        date(2019, 5, 2), country=Country.NORTHERN_IRELAND
-    )
-
-    assert publish_date == date(2019, 4, 8)
-
-
-# Reference election: local.herefordshire.ross-north.2019-06-06
-def test_publish_date_local_election_england():
-    publish_date = sopn_publish_date.local(date(2019, 6, 6), country=Country.ENGLAND)
-
-    assert publish_date == date(2019, 5, 10)
-
-
 # Reference election: mayor.liverpool-city-ca.2017-05-04
 def test_publish_date_mayor():
     publish_date = sopn_publish_date.mayor(date(2017, 5, 4))
