+++ uk_election_timetables/sopn.py
@@ -17,6 +17,7 @@ from uk_election_timetables.elections import (
     SeneddCymruElection,
     GreaterLondonAssemblyElection,
     NorthernIrelandAssemblyElection,
+    LocalElection,
 )
 
 
@@ -63,7 +64,7 @@ class StatementPublishDate(object):
             return self.election_id_lookup[election_type](poll_date)
 
         if election_type == "local":
-            return self.local(poll_date, country=country)
+            return LocalElection(poll_date, country).sopn_publish_date()
         elif election_type == "parl":
             return self.uk_parliament(poll_date, country=country)
 
@@ -134,31 +135,3 @@ class StatementPublishDate(object):
             ]
 
             return min(possible_dates)
-
-    def local(self, poll_date: date, country: Country):
-        """
-        Calculate the publish date for a local election.
-
-        This is set out in:
-
-         * `The Local Elections (Principal Areas) (England and Wales) (Amendment) Rules 2014 <https://www.legislation.gov.uk/uksi/2014/494/made>`_
-         * `The Local Elections (Northern Ireland) Order 2010 <https://www.legislation.gov.uk/uksi/2010/2977/schedule/1/part/4/made>`_
-         * `The Scottish Local Government Elections Order 2011 <https://www.legislation.gov.uk/ssi/2011/399/made>`_
-
-        :param poll_date: a datetime representing the date of the poll
-        :param country: the country in which the election is being run
-        :return: a datetime representing the expected publish date
-        """
-
-        country_specific_duration = {
-            Country.ENGLAND: 18,
-            Country.NORTHERN_IRELAND: 16,
-            Country.SCOTLAND: 23,
-            Country.WALES: 18,
-        }
-
-        days_prior = country_specific_duration[country]
-
-        return working_days_before(
-            poll_date, days_prior, self.calendar.from_country(country)
-        )
